requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pillow>=8.0.0
numpy>=1.19.2
gradio>=2.0.0
tqdm>=4.50.0 